fix: Reorganized Project Overview and Habit Data
